requirements.txt
Flask==2.2.3
spotipy==2.22.1
pandas==1.5.3
gunicorn==20.1.0
opencv_python_headless==4.7.0.72
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.11.0
scikit-learn==1.2.2
